Bug fix in Dot factory bitmaps when > 255 bytes in array. Added Access Point determination for OLED reporting of mode/IP address, now auto shows AP on icon Fixed latency issue when switching thermo/fixed mode in Screen3

+++ Arduino/BTCDieselHeater/FontTypes.h
@@ -35,7 +35,7 @@
 typedef struct  {
   uint8_t Width;                  // Char width in bits
   uint8_t Height;
-  uint8_t Offset;                 // Offset into bitmap array bytes)
+  uint16_t Offset;                 // Offset into bitmap array bytes)
 } FONT_CHAR_INFO;
 
 typedef struct  {
